requirements.txt
Django~=5.1.4
djangorestframework~=3.15.2
sqlparse~=0.5.3
python-dateutil~=2.9.0.post0
